requirements.txt
# Flux-Capacitor - dependências
# Core Web
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
jinja2==3.1.4

# Config
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# LLM + Orquestração
openai==1.51.0
langgraph==0.2.34
langchain-core==0.3.10
langchain-openai==0.2.2

# Observabilidade
langfuse==2.51.5

# DB
sqlalchemy==2.0.35
aiosqlite==0.20.0

# Utilidades
httpx==0.27.2
markdown==3.7
pyyaml==6.0.2

# Extração de conteúdo de uploads
pypdf==5.0.1
python-docx==1.1.2
openpyxl==3.1.5
Pillow==10.4.0
